pyproject.toml
[tool.poetry]
name = "plugin-runtime"
version = "0.1.0"
description = ""
authors = [
    "lujinghao <lujinghao@bytedance.com>",
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
pkginfo = "^1.11.2"
hiagent_plugin_sdk = { path = "hiagent-plugin-sdk" }
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
rq = "^2.0.0"
requests = "^2.32.3"
build = "^1.2.2.post1"
poetry-core = "^1.9.1"
aiofile = "^3.9.0"
minio = "^7.2.12"
boto3 = "^1.35.71"
entry-points-txt = "^0.2.0"
aiorwlock = "^1.5.0"
# uvicorn = "^0.32.0"
# hypercorn = "^0.17.3"
granian = "^1.7.6"
reretry = "^0.11.8"
jsonpath-ng = "^1.7.0"
openai = "^1.64.0"
sse-starlette = "^2.2.1"
aiohttp = "^3.11.13"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"
types-requests = "^2.32.0.20241016"
types-pyyaml = "^6.0.12.20240917"
httpx = "^0.27.2"
mypy = "^1.13.0"


[build-system]

requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
